--- .gitignore
@@ -1,4 +1,0 @@
-.idea
-vendor/
-.DS_Store
-composer.lock
