--- composer.json
@@ -1,37 +1,0 @@
-{
-    "name": "aiae/vito-copy-ssl",
-    "description": "Copy certificates from another site.",
-    "type": "library",
-    "license": "MIT",
-    "version": "1.0.1",
-    "authors": [
-        {
-            "name": "Daniel Dimitrov",
-            "email": "github@aiae.dev"
-        }
-    ],
-    "scripts": {
-        "test": "vendor/bin/phpunit"
-    },
-    "require": {
-        "php": "^8.4",
-        "illuminate/support": "^12.0"
-    },
-    "require-dev": {
-        "laravel/pint": "^1.10"
-    },
-    "autoload": {
-        "psr-4": {
-            "AiAe\\VitoCopySSL\\": "src"
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "AiAe\\VitoCopySSL\\VitoCopySSLPluginServiceProvider"
-            ]
-        }
-    },
-    "minimum-stability": "stable",
-    "prefer-stable": true
-}
